requirements.txt
numpy
matplotlib
pandas
cv2
Pillow